aruco dedector
hello this is a aruco dedector that measure the height and lengt. I must say that I just researched on the internet and combined what I found. Most of what I found gave errors. , so I fixed the problems that i see. it work but not perfetly, I just wanted to share.